Camera flicker fix
Fixed the issue in the camera switch script, method was called in the "Update" function using an "if statement" with only one condition that turned "True" causing the switch to happen on a loop each update; fixed by adding a "False bool CameraAlreadySwitched" and adding it as a "False condition" to the "if statement" and in the function body to turn "True" to make the "if statement" get called only once.

Deleted "TestScene" and made a new one.

hopefully the change in the script won't create a conflict. If it does, discard my changes and I'll explain further so the issue could be fix on Main Branch.